// src/types/evento.ts
export type Evento = {
  id: number;
  descricao: string;
  quantidade_maxima: number;
  quantidade_atual: number;
  visibilidade: boolean;
  rua: string;
  cidade: string;
  pais: string;
  cep: string;
  estado: string;
  id_usuario: string;
  id_categoria: string;
  categoria: string;
  usuario: string
}